% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crlapply.r
\name{crlapply}
\alias{crlapply}
\title{crlapply}
\usage{
crlapply(X, FUN, ..., checkpoint_file, checkpoint_freq = 1)
}
\arguments{
\item{X, FUN, ...}{Same as in \code{lapply()}.}

\item{checkpoint_file}{The checkpoint file.}

\item{checkpoint_freq}{The checkpoint frequency; a positive integer.}
}
\value{
A list.
}
\description{
An \code{lapply()}-like interface with automatic checkpoint/restart
functionality. Checkpoint/restart is a useful strategy for running expensive
functions in constrained environments (non-reliable hardware, restricted time
limits, etc).
}
\details{
The checkpoint file is cleaned up on successful completion of
\code{crlapply()}
}

