% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi_napply.r
\name{mpi_napply}
\alias{mpi_napply}
\title{mpi_napply}
\usage{
mpi_napply(n, FUN, ..., checkpoint_path = NULL, preschedule = TRUE)
}
\arguments{
\item{n}{A global, positive integer.}

\item{FUN}{Function to evaluate.}

\item{...}{Additional arguments passed to \code{FUN}.}

\item{checkpoint_path}{TODO}

\item{preschedule}{TODO}
}
\value{
TODO
}
\description{
An "n-apply" function.
}
\details{
TODO
}

