% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi_progress.r
\name{mpi_progress}
\alias{mpi_progress}
\title{mpi_progress}
\usage{
mpi_progress(checkpoint_path)
}
\arguments{
\item{checkpoint_path}{Path to the checkpoint files. Some extra restrictions here apply that do not
apply to the \code{mpi_*ply()} functions themselves. See the details section
for more information.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Show a progress bar for MPI tasks that use checkpointing. This is meant to be
called by an external process and not the "big job" itself. For example,
while the job is running, you might occasionally log in to the login node
and run \code{tasktools::mpi_progress("/path/to/checkpoints")}.
}
\details{
If all checkpoint files are in a single path, then you simply need to pass
that path (as a string) to the function. This is typical if you are using a
parallel file system. However, if the R processes use different paths for
the checkpoint files, then you can pass all of those paths as a character
vector. However, the calling process must be able to access all of those
paths (i.e. it can't be node-local).
}
