% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasktools-package.r
\docType{package}
\name{tasktools-package}
\alias{tasktools-package}
\title{tasktools-package}
\description{
Tools for task-based parallelism with MPI via pbdMPI.
}
\author{
Drew Schmidt
}
\keyword{package}
