% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasktools-package.r
\docType{package}
\name{tasktools-package}
\alias{tasktools-package}
\title{tasktools-package}
\description{
Tools for task-based parallelism with MPI via pbdMPI.
Several utilities are provided, each with an apply-like interface. There is
support for pre-scheduling tasks as well as a "load-balancing" mode that
farms them out from a central process. Finally, the functions support
checkpointing, which allows the task runs to be interrupted and restarted
from their last completed task.
}
\author{
Drew Schmidt
}
\keyword{package}
